XEEN: RU_T Split language constants Further sharpening ru translation
